requirements/requirements.txt
opencv-python-headless
nextcloud-api-wrapper
pytesseract
numpy
Pillow
fastapi
pydantic
uvicorn